requirement.txt
fastapi==0.110.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
httptools==0.6.1
lapx==0.5.5
librosa==0.10.1
mediapipe==0.10.11
pip-chill==1.0.3
pymysql==1.1.0
python-dotenv==1.0.1
tensorflow==2.16.1
ultralytics==8.1.25
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
